basic concepts R done

+++ OJ/Vjudge/F - Digits sum.c
@@ -4,12 +4,12 @@ int main()
     //may b previous solve it
     int n;
     scanf("%d",&n);
-    int dgt =0;
+    long long int dgt =0;
         while(n>0)
         {
             dgt = dgt + (n%10);
             n /=10;
         }
-    printf("%d\n",dgt);
+    printf("%lld\n",dgt);
     return 0;
 }
